requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
python-dotenv==0.19.2
gunicorn==20.1.0
requests==2.31.0
sqlite3
python-dateutil